Organized the sound systems directories, made sound source more similar to the texture class, and made additional comments in the SoundManager
